Fim da aula 3

+++ backend/src/controllers/IncidentController.js
@@ -1,12 +1,17 @@
 const connection = require("../database/connection");
 
 module.exports = {
-  async index(req, res) {
-    const { page = 1 } = req.query;
+  async index(request, response) {
+    const { page = 1 } = request.query;
+    // console.log(((page -1) * 5));
+
     const [count] = await connection("incidents").count();
 
+    response.header("X-Total-Count", count["count(*)"]);
+
+    //Paginação
     const incidents = await connection("incidents")
-      .join("ongs", "ong_id", "=", "incidents.ong_id")
+      .join("ongs", "ongs.id", "=", "incidents.ong_id")
       .limit(5)
       .offset((page - 1) * 5)
       .select([
@@ -14,38 +19,51 @@ module.exports = {
         "ongs.name",
         "ongs.email",
         "ongs.whatsapp",
-        "ongs.city",
         "ongs.uf"
       ]);
 
-    res.header("X-Total-Count", count["count(*)"] / 5);
-    res.header("X-Total-Pages", count["count(*)"]);
-    return res.json(incidents);
+    return response.json(incidents);
   },
-  async create(req, res) {
-    const { title, description, value } = req.body;
-    const ong_id = req.headers.authorization;
+  async create(request, response) {
+    //desestruturação
+    const { title, description, value } = request.body;
+    const ong_id = request.headers.authorization;
+
+    // const incident = await connection('incidents').insert({ //metho
     const [id] = await connection("incidents").insert({
       title,
       description,
       value,
       ong_id
     });
-    return res.json({ id });
+
+    // const id incident[0]; //metho
+    return response.json({ id });
   },
-  async delete(req, res) {
-    const { id } = req.params;
-    const ong_id = req.headers.authorization;
+  async delete(request, response) {
+    //desestruturação
+    const { id } = request.params;
+    //END - desestruturação
+    const ong_id = request.headers.authorization;
+
+    //Buscando incidents onde id = id e buscando apenas o valor ong_id para validação
     const incident = await connection("incidents")
       .where("id", id)
       .select("ong_id")
       .first();
-    if (incident.ong_id != ong_id) {
-      return res.status(401).json({ error: "Operação não permitida!" });
+
+    if (!incident) {
+      return response.json({ error: "Id inválid" });
     }
+
+    if (incident.ong_id != ong_id) {
+      return response.status(401).json({ error: "Operation not permitted." });
+    }
+
     await connection("incidents")
       .where("id", id)
       .delete();
-    return res.status(204).send();
+    //O send() é utilizado para enviar a resposta sem body
+    return response.status(204).send();
   }
 };
